+++ pkg/server/commands/list_objects_test.go
@@ -3,6 +3,7 @@ package commands
 import (
 	"context"
 	"testing"
+	"time"
 
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
@@ -226,4 +227,120 @@ func TestListObjectsDispatchCount(t *testing.T) {
 			require.Equal(t, test.expectedThrottlingValue > 0, resp.ResolutionMetadata.WasThrottled.Load())
 		})
 	}
+}
+
+func TestDoesNotUseCacheWhenHigherConsistencyEnabled(t *testing.T) {
+	ds := memory.New()
+	t.Cleanup(ds.Close)
+	ctx := storage.ContextWithRelationshipTupleReader(context.Background(), ds)
+	ctrl := gomock.NewController(t)
+	t.Cleanup(ctrl.Finish)
+	modelDsl := `model
+			schema 1.1
+
+			type user
+
+			type folder
+				relations
+					define viewer: [user] but not blocked
+					define blocked: [user]`
+	tuples := []string{
+		"folder:C#viewer@user:jon",
+		"folder:B#viewer@user:jon",
+		"folder:A#viewer@user:jon",
+	}
+
+	storeID, model := storagetest.BootstrapFGAStore(t, ds, modelDsl, tuples)
+	ts, err := typesystem.NewAndValidate(
+		context.Background(),
+		model,
+	)
+	require.NoError(t, err)
+
+	checkCache := storage.NewInMemoryLRUCache[*graph.ResolveCheckResponse]()
+	defer checkCache.Stop()
+
+	// Write an item to the cache that has an Allowed value of false for folder:A
+	req := &graph.ResolveCheckRequest{
+		StoreID: storeID,
+		TupleKey: &openfgav1.TupleKey{
+			User:     "user:jon",
+			Relation: "viewer",
+			Object:   "folder:A",
+		},
+	}
+	cacheKey, err := graph.CheckRequestCacheKey(req)
+	require.NoError(t, err)
+
+	checkCache.Set(cacheKey, &graph.ResolveCheckResponse{
+		Allowed: false,
+	}, 10*time.Second)
+
+	require.NoError(t, err)
+	ctx = typesystem.ContextWithTypesystem(ctx, ts)
+
+	checkResolver, checkResolverCloser := graph.NewOrderedCheckResolvers([]graph.CheckResolverOrderedBuilderOpt{
+		graph.WithCachedCheckResolverOpts(true, []graph.CachedCheckResolverOpt{
+			graph.WithEnabledConsistencyParams(true),
+			graph.WithExistingCache(checkCache),
+		}...),
+	}...).Build()
+	t.Cleanup(checkResolverCloser)
+
+	q, _ := NewListObjectsQuery(
+		ds,
+		checkResolver,
+	)
+
+	// Run a check with MINIMIZE_LATENCY that will use the cache we added with 2 tuples
+	resp, err := q.Execute(ctx, &openfgav1.ListObjectsRequest{
+		StoreId:     storeID,
+		Type:        "folder",
+		Relation:    "viewer",
+		User:        "user:jon",
+		Consistency: openfgav1.ConsistencyPreference_MINIMIZE_LATENCY,
+	})
+
+	require.NoError(t, err)
+	require.Len(t, resp.Objects, 2)
+
+	// Now run a check with HIGHER_CONSISTENCY that will evaluate against the known tuples and return 3 tuples
+	resp, err = q.Execute(ctx, &openfgav1.ListObjectsRequest{
+		StoreId:     storeID,
+		Type:        "folder",
+		Relation:    "viewer",
+		User:        "user:jon",
+		Consistency: openfgav1.ConsistencyPreference_HIGHER_CONSISTENCY,
+	})
+
+	require.NoError(t, err)
+	require.Len(t, resp.Objects, 3)
+
+	// Rerun check with MINIMIZE_LATENCY to ensure the cache was updated with the tuple we retrieved during the previous call
+	resp, err = q.Execute(ctx, &openfgav1.ListObjectsRequest{
+		StoreId:     storeID,
+		Type:        "folder",
+		Relation:    "viewer",
+		User:        "user:jon",
+		Consistency: openfgav1.ConsistencyPreference_MINIMIZE_LATENCY,
+	})
+
+	require.NoError(t, err)
+	require.Len(t, resp.Objects, 3)
+
+	// Now set the third item as `allowed: false` in the cache and run with `UNSPECIFIED`, it should use the cache and only return two item
+	checkCache.Set(cacheKey, &graph.ResolveCheckResponse{
+		Allowed: false,
+	}, 10*time.Second)
+
+	resp, err = q.Execute(ctx, &openfgav1.ListObjectsRequest{
+		StoreId:     storeID,
+		Type:        "folder",
+		Relation:    "viewer",
+		User:        "user:jon",
+		Consistency: openfgav1.ConsistencyPreference_UNSPECIFIED,
+	})
+
+	require.NoError(t, err)
+	require.Len(t, resp.Objects, 2)
 }